Enabling Capstone build for android-x86_64 arch variant
It looks to be a bug as currently the android.bp supports only arm,arm64 and x86. This change is added to ensure that we have a
valid build with android x86_64 arch variant for this capstone disassembler. We obseved crashes when we tried to
initialize this disassembler with x86_64 intel architecture at runtime. With the changes the capstone disassembler works as expected
for x86_64 variant as well

Test: Existing unit tests passes.

Change-Id: I50c6bf01e32cb4abfb8f319c1c380848a1e42af6
Signed-off-by: Priyanka Bose <priyanka.bose@intel.com>